ponchykBoy/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
celery[redis]==5.4.0
pillow==11.1.0
redis==5.2.1
djoser==2.3.1
phonenumbers==8.13.55
psycopg2-binary==2.9.10
flower==2.0.0